PersistenceTestKitBase is removed to correctly add ITestOutputHelper from xUnit
